% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum.plot.R
\name{cum.plot}
\alias{cum.plot}
\title{Cumulative frequency plot}
\usage{
cum.plot(file.or.object, legend.position = c(0.8, 0.7), incl.mito.genes = F,
  cum.freq.out.file.txt = "", file.pdf = "", file.html = "",
  color.numbers.to.use = "", biotype.cutoff = "")
}
\arguments{
\item{file.or.object}{Provide a .txt file or object where row.names are ensembl_gene_identifiers}

\item{legend.position}{Position of legend within plot. Defaults to c(0.8, 0.7)}

\item{incl.mito.genes}{Include mitochondrial genes in plot. Defaults to FALSE.}

\item{cum.freq.out.file.txt}{If specified, writes out the cumulative frequencies to a file.}

\item{file.pdf}{Where to store cum.plot. Defaults to getwd() and to filename "CPM.cumfreq.pdf".}

\item{file.html}{Where to store cum.plot. Defaults to getwd() and to filename "CPM.cumfreq.html".}

\item{color.numbers.to.use}{Specify colors to use for individual biotypes from the color palette brewer.pal.info[brewer.pal.info$category == 'qual',].}

\item{biotype.cutoff}{How many biotypes to represent in the plot. Defaults to 10.}
}
\examples{
See https://github.com/utnesp/cumulative.plot.R
}
\author{
Peter Utnes \email{utnesp@gmail.com}
}
